use flipnote lenny
In light of a recent DMCA regarding the redistribution of the fwTool 2.0 binary, we are now using Flipnote Lenny.
The DMCA's suggested resolution is to link directly to Lenny, however it's also been said that "these files" (presumably encompassing fwTool 2.0) "are not to be used on guide sites", so we are choosing to interpret the resolution presented in the DMCA as explicit permission to use Flipnote Lenny, acting as a contrary to the previous, much older statement.
fwTool 2.0 would never have been redistributed in the first place if the "flipnote lenny cannot be used on guide sites" restriction was not in effect.