Fix SD card root image
Apparently media.discordapp.com and cdn.discordapp.com have stopped
working unless viewed from within discord itself. Clicking "open
original" and removing everything after the ? also results in messages
like "This content is no longer available."

For now, just copy the images into the repo. The only instances of this
were for English and Spanish, so I have handled both of them, though
a French version of the image also exists, but is not used on the guide.

The better solution is to copy the HTML that is used to generate the
image into the repo itself, so it can be translated with the guide.
However, I don't have the web dev skills or knowledge to fix that.
If anyone can step up to do it, please do!!!!!!!!!!!!!!!!!!!!!!!!!!!!!