theme: separate user-facing text from code
Since somewhere between VuePress 2.0.0-rc.0 and rc.9, relative paths in
theme configuration stopped working (and apparently it isn't
considered a regression despite it having worked previously). [0]
Because of this, we can't slap a relative path and have Crowdin
replicate it to other locales anymore.

Unfortunately this means that the theme config can't be automated
anymore; however, we can still keep the translations flowing by
splitting user-facing text into a JSON and importing it for each
locale.

Along with splitting everything, we can variablize the locale and use
absolute paths via that variable, making it copy-pastable with just two
lines of change for each locale to enable the language.

It's a hack and I hate it, but this works and keeps everything in
relatively working condition so that we can update past VuePress
2.0.0-rc.0.

[0]: https://github.com/vuepress/ecosystem/issues/110